backend/requirements.txt
fastapi
uvicorn
pydantic
kafka-python
psycopg2-binary
